move classes to the new architecture
